cleanup the aftermath
