+++ src/graphics/cast_shadow.c
@@ -1,17 +1,5 @@
 #include "minirt.h"
 
-double	degrees(t_vector nrml, t_vector light_dir)
-{
-	double		tmp;
-	double		tmp2;
-	double		tmp3;
-
-	tmp = dot(nrml, light_dir);
-	tmp2 = sqrt(vec_length_squared(nrml));
-	tmp3 = sqrt(vec_length_squared(light_dir));
-	return (acos(tmp / (tmp2 * tmp3)) * 180.0 / M_PI);
-}
-
 static int	intercept_obj(t_map *map, t_hit hit, size_t i)
 {
 	t_vector	light_dir;
