+++ src/graphics/draw.c
@@ -6,25 +6,6 @@ void	draw_pixel(t_window *window, int x, int y, uint32_t color)
 		mlx_put_pixel(window->image, x, y, color);
 }
 
-uint32_t	write_color(double r, double g, double b, double a)
-{
-	int	ir;
-	int	ig;
-	int	ib;
-	int	ia;
-
-	ir = r * 0.999;
-	ig = g * 0.999;
-	ib = b * 0.999;
-	ia = a * 0.999;
-	return (ir << 24 | ig << 16 | ib << 8 | ia);
-}
-
-uint32_t	vec_to_color(t_vector vec)
-{
-	return (write_color(vec.x, vec.y, vec.z, 255.0));
-}
-
 void	draw(t_map *map)
 {
 	int			i;
